DialogueManager refactoring + moving CustomLine outside of the Helper class (with James' permission)
